# docs/openapi/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "AI Agent Productivity Platform - Open API",
        "contact": {},
        "version": "v0.0.1"
    },
    "basePath": "/openapi/v1",
    "paths": {
...
        "/chatflow/chat": {
            "post": {
                "description": "对话流OpenAPI",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "openapi"
                ],
                "summary": "对话流OpenAPI",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.OpenAPIChatflowChatRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/chatflow/conversation": {
            "post": {
                "description": "对话流创建对话OpenAPI",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "openapi"
                ],
                "summary": "对话流创建对话OpenAPI",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.OpenAPIChatflowCreateConversationRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.OpenAPIChatflowCreateConversationResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/chatflow/conversation/message/list": {
            "post": {
                "description": "对话流根据conversationId获取历史对话",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "openapi"
                ],
                "summary": "对话流根据conversationId获取历史对话",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "data",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/request.OpenAPIChatflowGetConversationMessageListRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.OpenAPIChatflowGetConversationMessageListResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/oauth/code/authorize": {
            "get": {
                "description": "授权码方式-获取授权码",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "openapi.OIDC"
                ],
                "summary": "授权码方式",
                "parameters": [
                    {
                        "type": "string",
                        "description": "备案ID",
                        "name": "client_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "重定向URI",
                        "name": "redirect_uri",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "响应类型",
                        "name": "response_type",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "权限范围",
                        "name": "scope",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "状态参数",
                        "name": "state",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "token",
                        "name": "jwt_token",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "302": {
                        "description": "重定向到指定URI",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/oauth/code/token": {
            "post": {
                "description": "授权码方式-获取Token",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "openapi.OIDC"
                ],
                "summary": "授权码方式",
                "parameters": [
                    {
                        "type": "string",
                        "description": "授权类型",
                        "name": "grant_type",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "授权码",
                        "name": "code",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Client ID",
                        "name": "client_id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "回调地址",
                        "name": "redirect_uri",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "备案密钥",
                        "name": "client_secret",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.OAuthTokenResponse"
                        }
                    }
                }
            }
        },
        "/oauth/code/token/refresh": {
            "post": {
                "description": "刷新令牌",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "openapi.OIDC"
                ],
                "summary": "刷新令牌",
                "parameters": [
                    {
                        "description": "RefreshToken",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.OAuthRefreshRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.OAuthRefreshTokenResponse"
                        }
                    }
                }
            }
        },
...
        "/oauth/login": {
            "get": {
                "description": "返回OAuth登录页面",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "openapi.OIDC"
                ],
                "summary": "OAuth登录授权",
                "parameters": [
                    {
                        "type": "string",
                        "description": "备案ID",
                        "name": "client_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "重定向URI",
                        "name": "redirect_uri",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "响应类型",
                        "name": "response_type",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "权限范围",
                        "name": "scope",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "状态参数",
                        "name": "state",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "302": {
                        "description": "重定向到指定URI",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/workflow/file/upload": {
            "post": {
                "description": "工作流OpenAPI文件上传",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "openapi"
                ],
                "summary": "工作流OpenAPI文件上传",
                "parameters": [
                    {
                        "type": "file",
                        "description": "文件",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/workflow/run": {
            "post": {
                "description": "工作流OpenAPI",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "openapi"
                ],
                "summary": "工作流OpenAPI",
                "parameters": [
                    {
                        "description": "工作流运行参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.OpenAPIWorkflowRunReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "request.OAuthRefreshRequest": {
            "type": "object",
            "required": [
                "grant_type"
            ],
            "properties": {
                "client_id": {
                    "type": "string"
                },
                "client_secret": {
                    "type": "string"
                },
                "grant_type": {
                    "type": "string"
                },
                "refresh_token": {
                    "type": "string"
                }
            }
        },
...
        "request.OpenAPIChatflowChatRequest": {
            "type": "object",
            "required": [
                "conversation_id",
                "query",
                "uuid"
            ],
            "properties": {
                "conversation_id": {
                    "type": "string"
                },
                "parameters": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "query": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
        "request.OpenAPIChatflowCreateConversationRequest": {
            "type": "object",
            "required": [
                "uuid"
            ],
            "properties": {
                "conversation_name": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
        "request.OpenAPIChatflowGetConversationMessageListRequest": {
            "type": "object",
            "required": [
                "conversation_id",
                "uuid"
            ],
            "properties": {
                "conversation_id": {
                    "type": "string"
                },
                "limit": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
...
        "request.OpenAPIWorkflowRunReq": {
            "type": "object",
            "required": [
                "uuid"
            ],
            "properties": {
                "parameters": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
...
        "response.OAuthRefreshTokenResponse": {
            "type": "object",
            "properties": {
                "access_token": {
                    "description": "访问令牌",
                    "type": "string"
                },
                "expires_at": {
                    "description": "token过期时间(毫秒时间戳)",
                    "type": "string"
                },
                "refresh_token": {
                    "description": "刷新令牌(可选)",
                    "type": "string"
                }
            }
        },
        "response.OAuthTokenResponse": {
            "type": "object",
            "properties": {
                "access_token": {
                    "description": "访问令牌",
                    "type": "string"
                },
                "expires_in": {
                    "description": "token过期时间(毫秒时间戳)",
                    "type": "integer"
                },
                "id_token": {
                    "description": "ID令牌",
                    "type": "string"
                },
                "refresh_token": {
                    "description": "刷新令牌",
                    "type": "string"
                },
                "scope": {
                    "description": "权限范围",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "token_type": {
                    "description": "令牌类型(bearer)",
                    "type": "string"
                }
            }
        },
...
        "response.OpenAPIChatflowCreateConversationResponse": {
            "type": "object",
            "properties": {
                "conversation_id": {
                    "type": "string"
                }
            }
        },
        "response.OpenAPIChatflowGetConversationMessageListResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.OpenMessageApi"
                    }
                },
                "first_id": {
                    "type": "integer"
                },
                "has_more": {
                    "type": "boolean"
                },
                "last_id": {
                    "type": "integer"
                }
            }
        },
...
        "response.OpenMessageApi": {
            "type": "object",
            "properties": {
                "bot_id": {
                    "type": "string",
                    "example": "0"
                },
                "chat_id": {
                    "type": "string",
                    "example": "0"
                },
                "content": {
                    "type": "string"
                },
                "content_type": {
                    "type": "string"
                },
                "conversation_id": {
                    "type": "string",
                    "example": "0"
                },
                "created_at": {
                    "type": "integer"
                },
                "id": {
                    "type": "string",
                    "example": "0"
                },
                "meta_data": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "reasoning_content": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "section_id": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "integer"
                }
            }
        },
        "response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
...
    }
}